requirements.txt
flask
flask-jwt-extended
flask-cors
python-dotenv
flask-sqlalchemy
flask-migrate
psycopg2-binary
werkzeug
langchain
